refactor: Remove unused function GetHelp and related code
The GetHelp function for returning help strings and its related code were
removed as they were no longer in use and deemed unnecessary. This cleanup
simplifies the codebase.
